Moved all sidenav content to a subfolder
