+++ app/scripts/sidenav/sidenav.jsx
@@ -3,7 +3,7 @@ import {Link} from 'react-router';
 
 import SidenavUserPanel from './sidenav-user-panel.jsx';
 
-require('../styles/sidenav.scss');
+require('../../styles/sidenav/sidenav.scss');
 
 /**
  * Sidenav component
