+++ app/styles/sidenav/sidenav-user-panel.scss
